pyproject.toml
[project]
name = "odin"
version = "0.1.0"
description = "AI-powered computer automation agent using vision LLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.0",
    "pillow>=12.1.0",
    "pyautogui>=0.9.54",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.14",
]
